+++ packages/media-library/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @templatical/media-library
 
+## 0.26.0
+
+### Patch Changes
+
+- Updated dependencies [753262e]
+  - @templatical/types@0.26.0
+  - @templatical/core@0.26.0
+
 ## 0.25.2
 
 ### Patch Changes